+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 -----
 
 *   New Features:
-    *   Fixed background color for screens using the compose theme
-        ([#432](https://github.com/Automattic/pocket-casts-android/pull/432)).
+    *   Added Tasker integration with "Play Filter" and "Control Playback" actions.
+        ([#415](https://github.com/Automattic/pocket-casts-android/pull/431)).
+    *   Import OPML from a URL
+        ([#482](https://github.com/Automattic/pocket-casts-android/pull/482)).
+    *   Redesign of the fullscreen player share option
+        ([#451](https://github.com/Automattic/pocket-casts-android/pull/451)).
 *   Bug Fixes:
     *   Fixed Help & Feedback buttons being hidden when using text zoom.
         ([#446](https://github.com/Automattic/pocket-casts-android/pull/446)).
@@ -11,6 +15,10 @@
         ([#461](https://github.com/Automattic/pocket-casts-android/pull/461)).
     *   Fixed When no podcasts were selected for a filter, change the chip to 'All Your Podcasts'
         ([#460](https://github.com/Automattic/pocket-casts-android/pull/460)).
+    *   Fixed background color for screens using the compose theme
+        ([#432](https://github.com/Automattic/pocket-casts-android/pull/432)).
+    *   Fixed full screen video player not closing the first time in landscape mode
+        ([#464](https://github.com/Automattic/pocket-casts-android/pull/464)).
 
 7.25
 -----